# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Textile's HTTP REST API Documentation",
        "title": "Textile REST API",
        "termsOfService": "https://github.com/b582q9/go-textile-sapien/blob/master/TERMS",
        "contact": {
            "name": "Textile",
            "url": "https://textile.io/",
            "email": "contact@textile.io"
        },
        "license": {
            "name": "MIT License",
            "url": "https://github.com/b582q9/go-textile-sapien/blob/master/LICENSE"
        },
        "version": "0"
    },
    "basePath": "/api/v0",
    "paths": {
...
        "/cafes": {
            "get": {
                "description": "List info about all active cafe sessions. Cafes are other peers on the network\nwho offer pinning, backup, and inbox services",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cafes"
                ],
                "summary": "List info about all active cafe sessions",
                "responses": {
                    "200": {
                        "description": "cafe sessions",
                        "schema": {
                            "$ref": "#/definitions/pb.CafeSessionList"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Registers with a cafe and saves an expiring service session token. An access\ntoken is required to register, and should be obtained separately from the target\nCafe",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cafes"
                ],
                "summary": "Register with a Cafe",
                "parameters": [
                    {
                        "type": "string",
                        "description": "cafe id",
                        "name": "X-Textile-Args",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "token=",
                        "description": "token: An access token supplied by the Cafe",
                        "name": "X-Textile-Opts",
                        "in": "header"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "cafe session",
                        "schema": {
                            "$ref": "#/definitions/pb.CafeSession"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "pb.Cafe": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "api": {
                    "type": "string"
                },
                "node": {
                    "type": "string"
                },
                "peer": {
                    "type": "string"
                },
                "protocol": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "pb.CafeSession": {
            "type": "object",
            "properties": {
                "access": {
                    "type": "string"
                },
                "cafe": {
                    "type": "object",
                    "$ref": "#/definitions/pb.Cafe"
                },
                "exp": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "refresh": {
                    "type": "string"
                },
                "rexp": {
                    "type": "string"
                },
                "subject": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "pb.CafeSessionList": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pb.CafeSession"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}